Changes to use a new configuration variable: CFG->slasharguments
If true, then display of user pictures or filenames will use the
method of providing arguments as "slash" arguments - this is much
better for caching, proxies, search engines etc.  Unfortunately
it doesn't seem to work on some PHP installations.

If false (the default), then a more compatible method is used
(ie usual way of passing parameters to scripts)