Modifications to allow Moodle files management page to use PclZip instead of external zip programs.
To use the internal PclZip, the config variables "zip" and "unzip"
must be left blank.

Code for PclZip integration originally from Robert Hetzel, though
I've expanded it and modified it a lot.

The zip files seem compatible with itself and Windows XP, but I had a
compatibility problem trying to unzip a PclZip file using Zip 2.3 on Linux
(directories are unpacked as empty files)