Fix for bug 1365:
This one was most insidious. Usually we get bugs that have to do with the
user/server timezone interaction, but by now I 've learned to be extra
careful about that. This one, however, only had to do with the server's
timezone and was 100% certain to happen if you are GMT minus something.
Considering that I 'm GMT+2, it went unnoticed. Sorry!