This fixes some rather serious bugs with how groups are used in group-oriented forums.
The zero group was used to define both the

   - ungrouped people
   - the posts that all groups should see

The groupid field in the forums is now a signed field and a new value -1 has
been introduced.

   0  --> ungrouped people
   -1 --> the posts that all groups should see

Things are now a lot more intuitive.  Teachers can post to individual
groups or to ALL groups at once.   Recent news display now makes sense etc.