Fix for bug 1494:
Calendar events in day view now show their start and end times. The only thing
I 'm not satisfied with is that they also show the full start and end dates,
even if it's not really needed (e.g. for a 15:00 -> 17:00 event). Still, it's
loads better than only showing the start time.