Bug 1848:
As per Gustav's recommendation, event names in Upcoming Events are now hyperlinked.
This required an internal API change, but as far as I can see only block_upcoming
used it the relevant functions so it's ok.
Also, substituded SITEID for the magic number 1 in calendar/lib.php as an aside.