Student and teacher enrolments now unique
Made student and teacher enrolments unique, enforced at the database level. This patch also removes a cleanup job in moodle's cron that removed duplicate enrolments.
The database changes are rolled as part of a version bump that we are *deferring* on MOODLE_14_STABLE.

Merge of patchset  arch-eduforge@catalyst.net.nz--2004/moodle--eduforge--1.3.3--patch-422