There is a new configuration variable called
   $THEME->standardsheets

that a theme can use to control how much of the
standard theme is included before the actual current
theme sheets.

When true, use all subsheets from "standard"

When not existent, use all subsheets from "standard" (this is to help old
themes work better out of the box)

When false, don't use standard at all

When an array of filenames, only use those (in that order).