Added Blackboard restore using XSLT.  Works with BB 5.5 and BB 6.0.
Completely untested by me but it LOOKS like it should work.  It's quite
standalone, and is basically a single function call during the restore
precheck which looks in the backup to see if it's a BB file, then
converts it using XSLT.

Apart from a little tidyup and renaming, I added a check so that it
should just silently fail if XSLT functions aren't available

Many thanks to Ziba Scott <ziba@linuxbox.com> who wrote it!