Refactored Assignment Module ============================
This new assignment module allows Plugin Assignment types.  This should
allow a whole lot of new stuff to easily be placed in Moodle without
making the modules list longer for no good reason.

This checkin is still a really rough version that needs work on it.
Normally I would want it better than this to check in, but I need
to collaborate with Shane on the remaining code and CVS is the best
way.

Give it a day or two to settle in.  :-)

This is the last thing I wanted to get into 1.5.  Getting close now!!

Cheers!
Martin