As per bug 2578:
Creating course and group events now requires the same privileges as posting
in a forum. For course events (including site course, i.e. site events), any
teacher can create one.

For group events, the teacher must either have editing rights in the course
or be a member of the group the event is for.