Fixed addslashes and stripslashes of responses. Now slashes are stripped in quiz_extract_responses, leading to an improvement in grading (before " a man's world" would have been graded as "a man\'s world").
In order for the responses to be written to the database correctly the function quiz_save_question_sessions adds slashes again before writing the data to the database (and before passing it to the questiontype method save_session_and_responses) and strips them again after all the work is done, thus leaving the responses fields unmodified for the subsequent code.

An additional benefit of this change is that answer fields that are automatically filled after a submit don't get more and more slashes added (e.g. ' used to become \' after the first and \\\' after the second commit).