Reverting toyomoyo's conversion of mod/quiz/view.php to roles because his changes made no sense.
Admittedly, this file is pretty incomprehensible and it is not clear at first what it is trying to do. (Even though it is more comprehensible now than it was a week ago.) I'll to a proper conversion of this file to RnP once I have checked in my other changes there - assuming I can actually get Moodle to work so I can test my changes.