some fixes for problem reported by Pierre Pichet.
Data set items form is now the opposite way round as the original was. This means that the add item section appears at the top of the form. Also if you press return while editing fields in the add item section the item will now be added. Default action for most browsers when a submit button is pressed is to act as if the first submit button in the form had been pressed.

Removed cancel buttons second and third pages of form too as they may not work as expected.