MDL-16606 "userdate inserts a space in front of date when chopping off the leading 0 on day of the month produced by %d"
extra space was used in front of %d so that we can detect the difference between a leading 0 and something like 10, 20, or 30. The extra space was removed along with any leading 0 but not if there was no leading zero.

Now removing extra space even if there is no leading zero.

Now applying fix to conditional branch for when using a time offset from GMT instead of server time.