blog MDLSITE-679 Some small patches to make blog RSS feeds work better.
a) You can define an array in config.php of exceptions to any rule which makes blogs not available to the outside world.  For example we use this on moodle.org with Helen's blog.   $CFG->openblogs = array(24152);

b) Author names are not atatched to every post in a 1-user feed.  Just the feed itself.

c) Links on user blogs are now more accurate and go right to the post.

d) Users are asked to log in if they aren't and guests don't have permission.