lib/phpmailer MDL-20701 - Convert to moodle_phpmailer class
Stage 2 - We now extend phpmailer and do our customisations in our subclass
rather than directly in the original file.

Previously we modified SetLanguage(), but I have not carried that change over
because I could not see code paths where this gets executed.(See bug)

I tested the header/content of the old class vs this one and think the changes
are all carried over. (If anyone has clever thoughts for how to unit test this
mailing out it would be useful, especially for upgrading phpmailer).