portfolio - MDL-19356 - add instance column to portfolio_tempdata
this is the first step in moving towards supporting multiple exports per session.

however, some plugins aren't going to be able to support this (eg box.net) so
we add the instance to the tempdata table, so that we can quickly check whether
we have an export of the same *type* of plugin in the session.