MDL-22015 Fixing get_string regression
The passed $a placeholder object can be some advanced object like $USER for example.
If such object contained other objects or arrays as properties, casting
to string in the loop throws error. This patch just skips all arrays and
objects when trying to find $a property value.
Thanks to Sam H. for spotting this.