auth cas: MDL-20029 upgrade phpCAS version to the latest available version
Also make sure phpCAS doesn't try to start a new PHP session. We have already
started our own and want to keep it :). In addition to it, it emitted a PHP
notice that could block the authentication process under certain
configurations.

And pretty much the same with the wantsurl and loginguest checks.